==> 2. StructureBasedResult.ipynb <==
# Penyatuan Data Spam Tipe 3 dan Hasil Perluasan Berbasis Struktur


import pandas as pd 
from nltk.tokenize import word_tokenize

Data Spam Tipe 3 

data_raw_s = pd.read_csv('Data/Output/1. Preprocessing/amz_end_prep_ag.csv', sep=',', encoding='latin-1')
data_raw_s['clean'] = data_raw_s['Body']

Data Hasil dari StructureBased.py

data3 = pd.read_csv('Data/Output/1. Preprocessing/Tag/non_pos.csv', sep=',', encoding='latin-1')
data3['clean'] = data3['Review']

# Penyatuan

data=pd.DataFrame(data_raw_s,columns=['clean'])
data=data.append(data3)
data=data.reset_index(drop=True)

# Eliminasi Review dengan Panjang Kurang dari 10 Kata

import re
for i in range(len(data['clean'])):
    sentences = re.split(r'(?<!\w\.\w.)(?<![A-Z][a-z]\.)(?<=\.|\?)\s',data['clean'][i])
    text = data['clean'][i].split()
    if len(text)<10:
        print(text)
        data.drop(i, inplace=True)

Menyimpan Data

data['clean'].to_csv(('Data/Output/1. Preprocessing/amz_end_prep_ag.csv'),header = 'clean', index=False)
