==> 3. SpamPreprocess.ipynb <==
# Spam Preprocessing

Tahap preposes yang akan dilakukan yaitu : Penghapusan simbol, link, kata-kata yang berulang, stop Word, dan Tokenisasi

import pandas as pd
from nltk.corpus import stopwords

# Fungsi Tokenisasi

* melakukan pemisahan kalimat menjadi list token kata
* menggunakan library nltk

def tokenization(stem_text):
    from nltk.tokenize import word_tokenize
    tokens = word_tokenize(str(stem_text))
    print(tokens)
    return tokens

# Fungsi Penghapusan Stopword

* melakukan penghapusan kata-kata yang termasuk stopword (kata umum yang sering muncul dan tidak memiliki makna, cont. 'the', that)

def stopword_removal(tokenized):
    clean = []
    for i in tokenized:
        if i not in stopwords.words('english'):
            clean.append(i)
    return clean

# Fungsi Preproses

* penggabungan penghapusan simbol/link/kata yang tidak diinginkan, penghapusan stopword, dan tokenisasi

def preprocess (text):
    text = text.str.lower()
    text = text.str.replace(r"htt\S+|pic\.\S+|pict\S+|www\S+|t\.\S+|bit\.\S+|goo\.\S+", " ") # remove http, https, pic, 
    text = text.str.replace(r"[^a-z]"," ") # remove non alphabetic
    text = text.str.replace(r"not\s+", "not_")
    text = text.str.replace(r"(.)\1{2,}", r"\1") # replace 3 repeated character in words into 1 character
    text = text.str.replace(r"(^| ).(( ).)*( |$)"," ") # remove 1 standalone character
    text = text.str.replace(r"\s\s+"," ") # remove multi space after regex
    text = text.str.lstrip()
    print(text.tail()) # preview data
    text = text.apply(tokenization) #tokenisasi
    text = text.apply(stopword_removal) #stopword removal
    return text

Preprocessing ini dilakukan pada data *Spam tipe 1*, *Spam tipe 3*, *Data dari Perluasan Berbasis Struktur* yang hasil akhirnya yang berupa toke akan digunakan untuk Proses Perluasan Berbasis Topik yang melibatkan Pemodelan Topik 


* Spam Tipe 3

data1 = pd.read_csv('Data/train_spam1 - juli.csv', sep=',', encoding='latin-1')
data1['clean']= preprocess(data1['Body'])
data1['clean'].to_csv(('Data/Output/1. Preprocessing/amz_spam1_token_ag.csv'),header = 'clean', index=False)


* Spam Tipe 1

#data3 = pd.read_csv('Data/train_data - juli.csv', sep=',', encoding='latin-1') 
#data3['clean']= preprocess(data2['Body'])
#data3['clean'].to_csv(('Data/Output/1. Preprocessing/amz_spam23_token_ag.csv'),header = 'clean', index=False)

* Dataset dari StructuredBased.py

#data4 = pd.read_csv('Data/Output/1. Preprocessing/Tag/non_pos.csv', sep=',', encoding='latin-1') #Data Perluasan Berbasis Struktur
#data4['clean']= preprocess(data3['Review'])
#data4['clean'].to_csv(('Data/Output/1. Preprocessing/amz_spam3_token_ag.csv'),header = 'clean', index=False)

* Data dari StructuredBasedResult.py

#data5 = pd.read_csv('Data/Output/1. Preprocessing/amz_end_prep_ag.csv', sep=',', encoding='latin-1') # Spam Tipe 3 + Data Perluasan Berbasis Struktur 
#data5['clean']= preprocess(data4['clean'])
#data6['clean'].to_csv(('Data/Output/1. Preprocessing/amz_spam323_token_ag.csv'),header = 'clean', index=False)
