==> 6. SequenceDesc.ipynb <==
# Pembentukan Sequence

* tahap ini diperlukan untuk penmebentukan data training dalam pembangkitan kata berbasis RNNs
* sequence dibentuk dari 50 token dan 1 token (target token).
* model Transformer tidak memerlukan bentuk pembagian token dengan target token karena berupa unsupervised learning

# Fungsi Load Doc 

* Membaca file txt

import string

# load doc into memory
def load_doc(filename):
    # open the file as read only
    file = open(filename, 'r')
    # read all text
    text = file.read()
    # close the file
    file.close()
    return text

# Fungsi *Cleaning Data*

* menghapus simbol, non alphabetic

# turn a doc into clean tokens
def clean_doc(doc):
    # replace '--' with a space ' '
    doc = doc.replace('--', ' ')
    # split into tokens by white space
    tokens = doc.split()
    # remove punctuation from each token
    table = str.maketrans('', '', string.punctuation)
    tokens = [w.translate(table) for w in tokens]
    # remove remaining tokens that are not alphabetic
    tokens = [word for word in tokens if word.isalpha()]
    # make lower case
    tokens = [word.lower() for word in tokens]
    return tokens

# Fungsi *Save Sequences*

* Menyimpan hasil pembentukan Sequence

# save tokens to file, one sequence per line
def save_doc(lines, filename):
    data = '\n'.join(lines)
    file = open(filename, 'w')
    file.write(data)
    file.close()

# Load Data

# load document
in_filename = 'Data/Output/5. Sequence/spam_seq_sel_ag3.txt'
doc = load_doc(in_filename)
# preview
print(doc[:200])

'''
# untuk file csv

import pandas as pd
doc = ""
datareview=pd.read_csv('Data/rev/train_spam1 - juli.csv', sep=',', encoding='latin-1') #data review dari 201 produk
for i in datareview['Body']:
    temp = []
    temp = ''.join(str(i)) +doc
    doc = temp
'''

# clean document
tokens = clean_doc(doc)
print(tokens[:20])
print('Total Tokens: %d' % len(tokens))
print('Unique Tokens: %d' % len(set(tokens)))


# Sequence Making

# organize into sequences of tokens
length = 50 + 1
sequences = list()
for i in range(length, len(tokens)):
    # select sequence of tokens
    seq = tokens[i-length:i]
    # convert into a line
    line = ' '.join(str(seq))
    # store
    sequences.append(line)
print('Total Sequences: %d' % len(sequences))

# Save Result

# save sequences to file
out_filename = 'Data/Output/5. Sequence/spam_seq_sel_done_ag3.txt'
save_doc(sequences, out_filename)
